--- assignment-2/assignment-2/Windows/AddVehiclesWindow.xaml
@@ -1,47 +1,0 @@
-﻿<Window x:Class="assignment_2.AddVehiclesWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:assignment_2"
-        mc:Ignorable="d"
-        Title="AddVehiclesWindow" Height="350" Width="525">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="279*"/>
-            <ColumnDefinition Width="238*"/>
-        </Grid.ColumnDefinitions>
-        <Button Content="Add" HorizontalAlignment="Left" Margin="23,288,0,0" VerticalAlignment="Top" Width="75" Click="Add_Click" Grid.Column="1" Height="22"/>
-        <Button Content="Back" HorizontalAlignment="Left" Margin="141,288,0,0" VerticalAlignment="Top" Width="75" Click="Back_Click" Grid.Column="1" Height="22"/>
-        <Label Content="Manufacturer:" HorizontalAlignment="Left" Margin="0,10,0,0" VerticalAlignment="Top" Height="26" Width="81"/>
-        <Label Content="MakeYear:" HorizontalAlignment="Left" Margin="0,36,0,0" VerticalAlignment="Top" Height="26" Width="81"/>
-        <Label Content="Registration:" HorizontalAlignment="Left" Margin="0,89,0,0" VerticalAlignment="Top" Height="26" Width="81" RenderTransformOrigin="0.48,1.473"/>
-        <Label Content="Total Km Travelled:" HorizontalAlignment="Left" Margin="0,115,0,0" VerticalAlignment="Top" Height="26" Width="110"/>
-        <Label Content="Total Services:" HorizontalAlignment="Left" Margin="0,146,0,0" VerticalAlignment="Top" Height="26" Width="90"/>
-        <Label Content="Revenue Recorded:" HorizontalAlignment="Left" Margin="0,208,0,0" VerticalAlignment="Top" Height="26" Width="115" RenderTransformOrigin="0.498,0.131"/>
-        <Label Content="Km Since Last Service:" HorizontalAlignment="Left" Margin="0,177,0,0" VerticalAlignment="Top" Height="26" Width="127" RenderTransformOrigin="0.596,-2.136"/>
-        <Label Content="Requires a Service:" HorizontalAlignment="Left" Margin="0,239,0,0" VerticalAlignment="Top" Height="26" Width="110" RenderTransformOrigin="0.465,0.989"/>
-        <TextBox Name="Manufacturer" HorizontalAlignment="Left" Height="23" Margin="139,12,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name ="makeYear" HorizontalAlignment="Left" Height="23" Margin="139,39,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <Label Content="Model:" HorizontalAlignment="Left" Margin="0,62,0,0" VerticalAlignment="Top" Width="81"/>
-        <TextBox Name ="Model" HorizontalAlignment="Left" Height="23" Margin="139,65,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name ="Registration" HorizontalAlignment="Left" Height="23" Margin="139,91,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name ="TotalKMS" HorizontalAlignment="Left" Height="23" Margin="139,118,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name="TServices" HorizontalAlignment="Left" Height="23" Margin="139,149,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name="KMSLS" HorizontalAlignment="Left" Height="23" Margin="139,177,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <TextBox Name="Revenue" HorizontalAlignment="Left" Height="23" Margin="139,207,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
-        <CheckBox Name="RS" Content="Yes" HorizontalAlignment="Left" Margin="139,244,0,0" VerticalAlignment="Top"/>
-        <Label Content="Name" Grid.Column="1" HorizontalAlignment="Left" Margin="0,10,0,0" VerticalAlignment="Top"/>
-        <Label Content="Name" Grid.Column="1" HorizontalAlignment="Left" Margin="0,88,0,0" VerticalAlignment="Top"/>
-        <Label Content="Name" Grid.Column="1" HorizontalAlignment="Left" Margin="0,62,0,0" VerticalAlignment="Top"/>
-        <Label Content="Number" Grid.Column="1" HorizontalAlignment="Left" Margin="0,36,0,0" VerticalAlignment="Top"/>
-        <Label Content="Number" Grid.Column="1" HorizontalAlignment="Left" Margin="0,115,0,0" VerticalAlignment="Top"/>
-        <Label Content="Number" Grid.Column="1" HorizontalAlignment="Left" Margin="0,146,0,0" VerticalAlignment="Top"/>
-        <Label Content="Number" Grid.Column="1" HorizontalAlignment="Left" Margin="0,177,0,0" VerticalAlignment="Top"/>
-        <Label Content="Number" Grid.Column="1" HorizontalAlignment="Left" Margin="0,208,0,0" VerticalAlignment="Top"/>
-        <TextBlock x:Name="Error" HorizontalAlignment="Left" Margin="10,294,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="259" Foreground="#FFBB2626"/>
-       
-
-
-    </Grid>
-</Window>
