+++ assignment-2/assignment-2/assignment-2.csproj
@@ -54,17 +54,10 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Windows\AddVehiclesWindow.xaml.cs">
-      <DependentUpon>AddVehiclesWindow.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Classes\FuelPurchase.cs" />
     <Compile Include="Classes\Journey.cs" />
     <Compile Include="Classes\Service.cs" />
     <Compile Include="Classes\Vehicles.cs" />
-    <Page Include="Windows\AddVehiclesWindow.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
